requirements.txt
sqlmodel==0.0.16
SQLAlchemy==2.0.29
fastapi==0.110.1
asyncmy===0.2.9
uvicorn[standard]===0.29.0
jinja2===3.1.3
fastapi-sessions===0.3.2
starlette==0.37.2
pyjwt===2.8.0
hypercorn===0.16.0
aiofiles==23.2.1